requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
autopep8==2.0.4
click==8.1.8
dill==0.3.9
docstring-to-markdown==0.16
exceptiongroup==1.2.2
fastapi==0.115.12
flake8==7.1.2
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.6.1
isort==6.0.1
jedi==0.19.2
mccabe==0.7.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.12.1
pydantic==2.11.2
pydantic_core==2.33.1
pydocstyle==6.3.0
pyflakes==3.2.0
pylint==3.3.6
python-dotenv==1.1.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
pytoolconfig==1.3.1
PyYAML==6.0.2
rope==1.13.0
sniffio==1.3.1
snowballstemmer==2.2.0
starlette==0.46.1
tomli==2.2.1
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.0
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
whatthepatch==1.0.7
yapf==0.43.0
zipp==3.21.0
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
SQLAlchemy==2.0.40
